improved Icon component: - reduce complexity by removing props file - using new theme-color structure
