+++ src/foundations/icon/Icon.constants.ts
@@ -8,18 +8,25 @@ const ICON_SIZES: TIconSizeToken[] = [8, 10, 12, 16, 20, 28, 32, 40, 52, 64, 104
 
 const ICON_GLYPHS: TIconGlyph[] = [
     'none',
-    ...IconGlyphs.map((glyph: string) => glyph as TIconGlyph),
+    ...IconGlyphs.filter((glyph) => !glyph.includes('BRKN')).map(
+        (glyph: string) => glyph as TIconGlyph
+    ),
 ];
 
 const DEFAULT_ICON_GLYPH: TIconGlyph = 'mattermost';
 
 const ICON_COLORS: TIconColor[] = [
     'inherit',
+    'navigation',
+    'content',
+    'mention',
     'primary',
     'secondary',
+    'tertiary',
     'alert',
     'warning',
     'success',
+    'info',
 ];
 
 const DEFAULT_ICON_COLOR: TIconColor = 'inherit';
