--- src/foundations/icon/Icon.props.ts
@@ -1,38 +1,0 @@
-import type React from 'react';
-
-import type { TIconColor, TIconGlyph, TIconSizeToken } from './Icon.types';
-
-type PIcon = {
-    /**
-     * the size the icon is rendered with
-     * @default 20
-     */
-    size?: TIconSizeToken;
-    /**
-     * the color token the Icon should be rendered with.
-     * when not passed a value it will inherit the color.
-     * @default 'inherit'
-     */
-    color?: TIconColor;
-    /**
-     * the icon-glyph that is being rendered
-     */
-    glyph?: TIconGlyph;
-    /**
-     * add an aria-label for a11y
-     */
-    ariaLabel?: string;
-    /**
-     * custom className
-     */
-    className?: string;
-    /**
-     * onClick handler
-     */
-    onClick?: React.MouseEventHandler;
-};
-
-export type PIconRoot = Required<Pick<PIcon, 'size' | 'color'>> &
-    Pick<PIcon, 'ariaLabel' | 'className'>;
-
-export default PIcon;
