+++ src/foundations/icon/Icon.stories.mdx
@@ -106,7 +106,7 @@ not the font size for the icon font.
                     type: 'text',
                 },
             },
-            ...Utils.hideComponentProperties(['className']),
+            ...Utils.hideComponentProperties(['className', 'onClick']),
         }}
     >
         {(args) => <Icon {...args} />}
