+++ src/foundations/icon/Icon.tsx
@@ -4,7 +4,6 @@ import type { NonBrokenIconGlyphTypes } from '@mattermost/compass-icons/IconGlyp
 
 import { Utils } from '../../shared';
 
-import type PIcon from './Icon.props';
 import IconRoot from './Icon.root';
 import {
     DEFAULT_ICON_COLOR,
@@ -12,6 +11,37 @@ import {
     DEFAULT_ICON_SIZE,
     ICON_GLYPHS,
 } from './Icon.constants';
+import type { TIconColor, TIconGlyph, TIconSizeToken } from './Icon.types';
+
+type PIcon = {
+    /**
+     * the size the icon is rendered with
+     * @default 20
+     */
+    size?: TIconSizeToken;
+    /**
+     * the color token the Icon should be rendered with.
+     * when not passed a value it will inherit the color.
+     * @default 'inherit'
+     */
+    color?: TIconColor;
+    /**
+     * the icon-glyph that is being rendered
+     */
+    glyph?: TIconGlyph;
+    /**
+     * add an aria-label for a11y
+     */
+    ariaLabel?: string;
+    /**
+     * custom className
+     */
+    className?: string;
+    /**
+     * onClick handler
+     */
+    onClick?: React.MouseEventHandler;
+};
 
 const Icon = (props: PIcon): JSX.Element => {
     const {
@@ -28,6 +58,7 @@ const Icon = (props: PIcon): JSX.Element => {
         )}.`
     );
 
+    // we have a "placeholder" icon ('none') that does not render an icon at all
     if (glyph === 'none') {
         return <IconRoot size={size} color={'inherit'} {...rest} />;
     }
@@ -41,4 +72,5 @@ const Icon = (props: PIcon): JSX.Element => {
     );
 };
 
+export type { PIcon };
 export default Icon;
