+++ src/foundations/icon/Icon.types.ts
@@ -1,11 +1,11 @@
-import type IconGlyphs from '@mattermost/compass-icons/IconGlyphs';
+import type { NonBrokenIconGlyphTypes } from '@mattermost/compass-icons/IconGlyphs';
 
-import type { TTHemeColors } from '../../utilities/theme';
+import type { TNewTheme } from '../../utilities/theme';
 
-type TIconGlyph = 'none' | typeof IconGlyphs[number];
+type TIconGlyph = 'none' | NonBrokenIconGlyphTypes;
 
 type TIconSizeToken = 8 | 10 | 12 | 16 | 20 | 28 | 32 | 40 | 52 | 64 | 104;
 
-type TIconColor = 'inherit' | keyof TTHemeColors;
+type TIconColor = 'inherit' | keyof TNewTheme['palettes'];
 
 export type { TIconGlyph, TIconSizeToken, TIconColor };
