+++ src/foundations/icon/index.ts
@@ -1,5 +1,5 @@
+export * from './Icon';
 export * from './Icon.types';
-export * from './Icon.props';
 export * from './Icon.constants';
 
 export { default } from './Icon';
